new URI to notify new CBR message displayed to users
a new URI content://cellbroadcasts/displayed used to update
entries in cellbroadcastservice database that the message was
displayed to users either by alert dialog or notifications.
Also, contentObervers can listen for this URI for any new displayed
alert messages. The notified URI will be in the format of
content://cellbroadcasts/displayed/<row_number>

Bug: 163430905
Test: Manual test
Change-Id: Ic8bb59f4fc20cd49af6aeb92a79086e2b4391067